requirements.txt
anyio==4.4.0
black==24.8.0
certifi==2024.7.4
click==8.1.7
exceptiongroup==1.2.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
logging==0.4.9.6
mccabe==0.7.0
mypy-extensions==1.0.0
notion-client==2.2.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
-e git+ssh://git@github.com/misha-wu/notion_takehome.git@d91a3a6f71d7fb237f8212c1a7239e5709ecbe1b#egg=sonnet
tomli==2.0.1
typing_extensions==4.12.2
zipp==3.20.0
